+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgrade Go version to 1.21 ([#2](https://github.com/gopxl/beep/pull/2))
 - Upgrade Oto version to 3.1 ([#3](https://github.com/gopxl/beep/pull/3))
+- Panic when `Resampler` is given a ratio of `Inf` or `NaN`. ([#120](https://github.com/gopxl/beep/pull/120))
 
 ## [v1.0.0] 2023-10-07
 - Forked [faiface/beep](https://github.com/faiface/beep) to [gopxl/beep](https://github.com/gopxl/beep).